requirements.txt
fastapi
uvicorn
python-telegram-bot
langchain
langchain-community
langchain_core
langchain_openai
ollama
python-dotenv
httpx
duckduckgo-search
ddgs
supabase
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
